requirements.txt
pandas
python-dotenv
flask
flask-cors
requests
scikit-learn
statsmodels
